IssueID:1580:HaaS200 flash driver for new driver framework
[Detail]
HaaS200 flash driver for new driver framework.

[Verified Cases]
Build Pass: helloworld_demo@haas200
Test Pass: helloworld_demo@haas200
